backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.10.4
boto3==1.35.90
python-dotenv==1.0.1

